script: Add scripts for deployment & operation

+++ src/errors/Error.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: UNLICENSED
+// SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
 import {RevealType} from "src/types/GlobalEnum.sol";
@@ -15,21 +15,21 @@ library ConfigError {
      * @param mintPrice_ The mint price.
      * @param msgValue_ The message value.
      */
-    error InsufficientPayment(uint mintPrice_, uint msgValue_);
+    error InsufficientPayment(uint256 mintPrice_, uint256 msgValue_);
 
     /**
      * @dev When setting the mint start block, it should be greater than the current block.
      * @param mintStartBlock_ The mint start block.
      * @param currentBlock_ The current block.
      */
-    error MintStartBlockShouldBeGreaterThanCurrentBlock(uint mintStartBlock_, uint currentBlock_);
+    error MintStartBlockShouldBeGreaterThanCurrentBlock(uint256 mintStartBlock_, uint256 currentBlock_);
 
     /**
      * @dev When setting the reveal start block, it should be greater than the mint start block.
      * @param mintStartBlock_ The mint start block.
      * @param revealStartBlock_ The reveal start block.
      */
-    error MintStartShouldEarlierThanRevealStart(uint mintStartBlock_, uint revealStartBlock_);
+    error MintStartShouldEarlierThanRevealStart(uint256 mintStartBlock_, uint256 revealStartBlock_);
 
     /**
      * @dev The configuration has already been initialized.
@@ -76,13 +76,13 @@ library RevealError {
      * @dev The token ID has already been revealed.
      * @param tokenId_ The token ID which has already been revealed.
      */
-    error RevealProcessAlreadyStarted(uint tokenId_);
+    error RevealProcessAlreadyStarted(uint256 tokenId_);
 
     /**
      * @dev The token ID has not in the reveal process.
      * @param tokenId_ The token ID which has not in the reveal process.
      */
-    error RevealProcessNotInProgress(uint tokenId_);
+    error RevealProcessNotInProgress(uint256 tokenId_);
 }
 
 // Collection Error
